+++ CHANGELOG.md
@@ -9,11 +9,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- 
+- Generated files are displayed under the underneath the response
 
 ### Changed
 
-- 
+- Summaries properly under files (disabled by default)
 
 ### Removed